Movement script with camera jitter
The movement is working perfectly, but there's a bug where the camera jitters.
This mostly happens when you jump.
It happens because we're moving at a different time than when the camera is rotating. To fix this, we should change the move direction to use a different camera forward instead of the normal one we're using. This controller doesn't use the orientation transform most tutorials do, because it's not necessary with the fix we're going to make.